changed switching api from 0..1 to 0..100. Added migration for scenes

+++ ios/WatchTower Extension/Logic/BluenetManager.swift
@@ -165,7 +165,7 @@ class BluenetManager {
     }
 
   
-    public func switchStoneBroadcast(_ referenceId: String, stoneId: UInt8, _ newState: Float) {
+    public func switchStoneBroadcast(_ referenceId: String, stoneId: UInt8, _ newState: UInt8) {
      _ = self.bluenet.broadcast.multiSwitch(referenceId: referenceId, stoneId: stoneId, switchState: newState)
     }
